+++ sdk/src/store.rs
@@ -13,10 +13,13 @@
 
 use std::{
     collections::HashMap,
-    io::{Cursor, Read, Seek, SeekFrom},
+    io::{Cursor, Read, Seek},
 };
 #[cfg(feature = "file_io")]
-use std::{fs, path::Path};
+use std::{
+    fs,
+    path::{Path, PathBuf},
+};
 
 use async_generic::async_generic;
 use async_recursion::async_recursion;
@@ -61,6 +64,7 @@ use crate::{
     trust_handler::TrustHandlerConfig,
     utils::{
         hash_utils::{hash_sha256, HashRange},
+        io_utils::stream_len,
         patch::patch_bytes,
     },
     validation_status, AsyncSigner, RemoteSigner, Signer,
@@ -1395,7 +1399,8 @@ impl Store {
     }
 
     // wake the ingredients and validate
-    #[async_recursion(?Send)]
+    #[cfg_attr(target_arch = "wasm32", async_recursion(?Send))]
+    #[cfg_attr(not(target_arch = "wasm32"), async_recursion)]
     async fn ingredient_checks_async(
         store: &Store,
         claim: &Claim,
@@ -1441,7 +1446,7 @@ impl Store {
                     }
 
                     let check_ingredient_trust: bool =
-                        crate::settings::get_settings_value("verify.check_ingredient_trust")?;
+                        get_settings_value("verify.check_ingredient_trust")?;
 
                     // verify the ingredient claim
                     Claim::verify_claim_async(
@@ -1584,7 +1589,7 @@ impl Store {
             return Ok(out);
         }
 
-        let stream_len = stream.seek(SeekFrom::End(0))?;
+        let stream_len = stream_len(stream)?;
         stream.rewind()?;
 
         let mut hashes: Vec<DataHash> = Vec::new();
@@ -1649,17 +1654,14 @@ impl Store {
         Ok(hashes)
     }
 
-    fn generate_bmff_data_hashes_for_stream(
+    fn generate_bmff_data_hash_for_stream(
         asset_stream: &mut dyn CAIRead,
         alg: &str,
         calc_hashes: bool,
-    ) -> Result<Vec<BmffHash>> {
-        use serde_bytes::ByteBuf;
-
+        flat_fragmented_w_merkle: bool,
+    ) -> Result<BmffHash> {
         // The spec has mandatory BMFF exclusion ranges for certain atoms.
         // The function makes sure those are included.
-
-        let mut hashes: Vec<BmffHash> = Vec::new();
 
         let mut dh = BmffHash::new("jumbf manifest", alg, None);
         let exclusions = dh.exclusions_mut();
@@ -1680,6 +1682,11 @@ impl Store {
         let ftyp = ExclusionsMap::new("/ftyp".to_owned());
         exclusions.push(ftyp);
 
+        // /mfra/ exclusion
+        let mfra = ExclusionsMap::new("/mfra".to_owned());
+        exclusions.push(mfra);
+
+        /*  no longer mandatory
         // meta/iloc exclusion
         let iloc = ExclusionsMap::new("/meta/iloc".to_owned());
         exclusions.push(iloc);
@@ -1729,19 +1736,19 @@ impl Store {
         trun.subset = Some(subset_trun_vec);
         trun.flags = Some(ByteBuf::from([1, 0, 0]));
         exclusions.push(trun);
-
-        // V2 exclusions
-        /*  Enable this when we support Merkle trees and fragmented MP4
-        // /mdat exclusion
-        let mut mdat = ExclusionsMap::new("/mdat".to_owned());
-        let subset_mdat = SubsetMap {
-            offset: 16,
-            length: 0,
-        };
-        let subset_mdat_vec = vec![subset_mdat];
-        mdat.subset = Some(subset_mdat_vec);
-        exclusions.push(mdat);
         */
+
+        // enable flat flat files with Merkle trees
+        if flat_fragmented_w_merkle {
+            let mut mdat = ExclusionsMap::new("/mdat".to_owned());
+            let subset_mdat = SubsetMap {
+                offset: 16,
+                length: 0,
+            };
+            let subset_mdat_vec = vec![subset_mdat];
+            mdat.subset = Some(subset_mdat_vec);
+            exclusions.push(mdat);
+        }
 
         if calc_hashes {
             dh.gen_hash_from_stream(asset_stream)?;
@@ -1753,9 +1760,8 @@ impl Store {
                 _ => return Err(Error::UnsupportedType),
             }
         }
-        hashes.push(dh);
 
-        Ok(hashes)
+        Ok(dh)
     }
 
     // move or copy data from source to dest
@@ -2049,6 +2055,113 @@ impl Store {
             }
         }
         Err(Error::UnsupportedType)
+    }
+
+    #[cfg(feature = "file_io")]
+    fn start_save_bmff_fragmented(
+        &mut self,
+        asset_path: &Path,
+        fragments: &Vec<std::path::PathBuf>,
+        output_dir: &Path,
+        reserve_size: usize,
+    ) -> Result<Vec<u8>> {
+        // get the provenance claim changing mutability
+        let pc = self.provenance_claim_mut().ok_or(Error::ClaimEncoding)?;
+        pc.clear_data(); // clear since we are reusing an existing claim
+
+        let output_filename = asset_path.file_name().ok_or(Error::NotFound)?;
+        let dest_path = output_dir.join(output_filename);
+
+        let mut data;
+
+        // 2) Get hash ranges if needed
+        let mut asset_stream = std::fs::File::open(asset_path)?;
+        let mut bmff_hash =
+            Store::generate_bmff_data_hash_for_stream(&mut asset_stream, pc.alg(), false, false)?;
+        bmff_hash.clear_hash();
+
+        // generate fragments and produce Merkle tree
+        bmff_hash.add_merkle_for_fragmented(
+            pc.alg(),
+            asset_path,
+            fragments,
+            output_dir,
+            1,
+            None,
+        )?;
+
+        // add in the BMFF assertion
+        pc.add_assertion(&bmff_hash)?;
+
+        // 3) Generate in memory CAI jumbf block
+        // and write preliminary jumbf store to file
+        // source and dest the same so save_jumbf_to_file will use the same file since we have already cloned
+        data = self.to_jumbf_internal(reserve_size)?;
+        let jumbf_size = data.len();
+        save_jumbf_to_file(&data, &dest_path, Some(&dest_path))?;
+
+        // generate actual hash values
+        let pc = self.provenance_claim_mut().ok_or(Error::ClaimEncoding)?; // reborrow to change mutability
+
+        let bmff_hashes = pc.bmff_hash_assertions();
+
+        if !bmff_hashes.is_empty() {
+            let mut bmff_hash = BmffHash::from_assertion(bmff_hashes[0])?;
+            bmff_hash.update_fragmented_inithash(&dest_path)?;
+            pc.update_bmff_hash(bmff_hash)?;
+        }
+
+        // regenerate the jumbf because the cbor changed
+        data = self.to_jumbf_internal(reserve_size)?;
+        if jumbf_size != data.len() {
+            return Err(Error::JumbfCreationError);
+        }
+
+        Ok(data) // return JUMBF data
+    }
+
+    /// Embed the claims store as jumbf into fragmented assets.
+    #[cfg(feature = "file_io")]
+    pub fn save_to_bmff_fragmented(
+        &mut self,
+        asset_path: &Path,
+        fragments: &Vec<std::path::PathBuf>,
+        output_path: &Path,
+        signer: &dyn Signer,
+    ) -> Result<()> {
+        match get_supported_file_extension(asset_path) {
+            Some(ext) => {
+                if !is_bmff_format(&ext) {
+                    return Err(Error::UnsupportedType);
+                }
+            }
+            None => return Err(Error::UnsupportedType),
+        }
+
+        let mut validation_log = OneShotStatusTracker::new();
+        let jumbf = self.to_jumbf(signer)?;
+
+        // use temp store so mulitple calls will work (the Store is not finalized this way)
+        let mut temp_store = Store::from_jumbf(&jumbf, &mut validation_log)?;
+
+        let jumbf_bytes = temp_store.start_save_bmff_fragmented(
+            asset_path,
+            fragments,
+            output_path,
+            signer.reserve_size(),
+        )?;
+
+        let pc = temp_store.provenance_claim().ok_or(Error::ClaimEncoding)?;
+        let sig = temp_store.sign_claim(pc, signer, signer.reserve_size())?;
+        let sig_placeholder = Store::sign_claim_placeholder(pc, signer.reserve_size());
+
+        let output_filename = asset_path.file_name().ok_or(Error::NotFound)?;
+        let dest_path = output_path.join(output_filename);
+
+        match temp_store.finish_save(jumbf_bytes, &dest_path, sig, &sig_placeholder) {
+            Ok(_) => Ok(()),
+            Err(e) => Err(e),
+        }
     }
 
     /// Embed the claims store as jumbf into a stream. Updates XMP with provenance record.
@@ -2603,14 +2716,13 @@ impl Store {
             // 2) Get hash ranges if needed, do not generate for update manifests
             if !pc.update_manifest() {
                 intermediate_stream.rewind()?;
-                let bmff_hashes = Store::generate_bmff_data_hashes_for_stream(
+                let bmff_hash = Store::generate_bmff_data_hash_for_stream(
                     &mut intermediate_stream,
                     pc.alg(),
                     false,
+                    false,
                 )?;
-                for hash in bmff_hashes {
-                    pc.add_assertion(&hash)?;
-                }
+                pc.add_assertion(&bmff_hash)?;
             }
 
             // 3) Generate in memory CAI jumbf block
@@ -2854,11 +2966,9 @@ impl Store {
             // 2) Get hash ranges if needed, do not generate for update manifests
             if !pc.update_manifest() {
                 let mut file = std::fs::File::open(asset_path)?;
-                let bmff_hashes =
-                    Store::generate_bmff_data_hashes_for_stream(&mut file, pc.alg(), false)?;
-                for hash in bmff_hashes {
-                    pc.add_assertion(&hash)?;
-                }
+                let bmff_hash =
+                    Store::generate_bmff_data_hash_for_stream(&mut file, pc.alg(), false, false)?;
+                pc.add_assertion(&bmff_hash)?;
             }
 
             // 3) Generate in memory CAI jumbf block
@@ -3190,11 +3300,10 @@ impl Store {
 
                 Ok(store)
             })
-            .map_err(|e| {
+            .inspect_err(|e| {
                 validation_log.log_silent(
-                    log_item!("asset", "error loading file", "load_from_asset").set_error(&e),
+                    log_item!("asset", "error loading file", "load_from_asset").set_error(e),
                 );
-                e
             })
     }
 
@@ -3204,11 +3313,10 @@ impl Store {
         validation_log: &mut impl StatusTracker,
     ) -> Result<Store> {
         // load jumbf if available
-        Self::load_cai_from_memory(asset_type, data, validation_log).map_err(|e| {
+        Self::load_cai_from_memory(asset_type, data, validation_log).inspect_err(|e| {
             validation_log.log_silent(
-                log_item!("asset", "error loading asset", "get_store_from_memory").set_error(&e),
+                log_item!("asset", "error loading asset", "get_store_from_memory").set_error(e),
             );
-            e
         })
     }
 
@@ -3289,6 +3397,38 @@ impl Store {
         }
 
         Ok(store)
+    }
+
+    /// Load Store from a init and fragments
+    /// asset_type: asset extension or mime type
+    /// init_segment: reader for the file containing the initialization segments
+    /// fragments: list of paths to the fragments to verify
+    /// verify: if true will run verification checks when loading, all fragments must verify for Ok status
+    /// validation_log: If present all found errors are logged and returned, otherwise first error causes exit and is returned
+    #[cfg(feature = "file_io")]
+    pub fn load_from_file_and_fragments(
+        asset_type: &str,
+        init_segment: &mut dyn CAIRead,
+        fragments: &Vec<PathBuf>,
+        verify: bool,
+        validation_log: &mut impl StatusTracker,
+    ) -> Result<Store> {
+        let mut init_seg_data = Vec::new();
+        init_segment.read_to_end(&mut init_seg_data)?;
+
+        Store::get_store_from_memory(asset_type, &init_seg_data, validation_log).and_then(|store| {
+            // verify the store
+            if verify {
+                // verify store and claims
+                Store::verify_store(
+                    &store,
+                    &mut ClaimAssetData::StreamFragments(init_segment, fragments, asset_type),
+                    validation_log,
+                )?;
+            }
+
+            Ok(store)
+        })
     }
 
     /// Load Store from a in-memory asset
@@ -5543,6 +5683,8 @@ pub mod tests {
     #[cfg(feature = "file_io")]
     async fn test_datahash_embeddable_manifest_async() {
         // test adding to actual image
+
+        use std::io::SeekFrom;
         let ap = fixture_path("cloud.jpg");
 
         // Do we generate JUMBF?
@@ -5610,6 +5752,8 @@ pub mod tests {
     #[cfg(feature = "file_io")]
     fn test_datahash_embeddable_manifest() {
         // test adding to actual image
+
+        use std::io::SeekFrom;
         let ap = fixture_path("cloud.jpg");
 
         // Do we generate JUMBF?
@@ -5677,6 +5821,8 @@ pub mod tests {
     #[cfg(feature = "file_io")]
     fn test_datahash_embeddable_manifest_user_hashed() {
         // test adding to actual image
+
+        use std::io::SeekFrom;
         let ap = fixture_path("cloud.jpg");
 
         let mut hasher = Hasher::SHA256(Sha256::new());
@@ -5854,5 +6000,101 @@ pub mod tests {
 
         let errors = report_split_errors(report.get_log_mut());
         assert!(errors.is_empty());
+    }
+
+    #[test]
+    #[cfg(feature = "file_io")]
+    fn test_fragmented_jumbf_generation() {
+        // test adding to actual image
+
+        let tempdir = tempdir().expect("temp dir");
+        let output_path = tempdir.into_path();
+
+        // search folders for init segments
+        for init in glob::glob(
+            fixture_path("bunny/**/BigBuckBunny_2s_init.mp4")
+                .to_str()
+                .unwrap(),
+        )
+        .unwrap()
+        {
+            match init {
+                Ok(p) => {
+                    let mut fragments = Vec::new();
+                    let init_dir = p.parent().unwrap();
+                    let seg_glob = init_dir.join("BigBuckBunny_2s*.m4s"); // segment match pattern
+
+                    // grab the fragments that go with this init segment
+                    for seg in glob::glob(seg_glob.to_str().unwrap()).unwrap().flatten() {
+                        fragments.push(seg);
+                    }
+
+                    // Create claims store.
+                    let mut store = Store::new();
+
+                    // Create a new claim.
+                    let claim = create_test_claim().unwrap();
+                    store.commit_claim(claim).unwrap();
+
+                    // Do we generate JUMBF?
+                    let signer = temp_signer();
+
+                    // add manifest based on
+                    let new_output_path = output_path.join(init_dir.file_name().unwrap());
+                    store
+                        .save_to_bmff_fragmented(
+                            p.as_path(),
+                            &fragments,
+                            new_output_path.as_path(),
+                            signer.as_ref(),
+                        )
+                        .unwrap();
+
+                    // verify the fragments
+                    let output_init = new_output_path.join(p.file_name().unwrap());
+                    let init_stream = std::fs::read(output_init).unwrap();
+
+                    for entry in &fragments {
+                        let file_path = new_output_path.join(entry.file_name().unwrap());
+
+                        let mut validation_log = DetailedStatusTracker::new();
+
+                        let fragment_stream = std::fs::read(&file_path).unwrap();
+                        let _manifest = Store::load_fragment_from_memory(
+                            "mp4",
+                            &init_stream,
+                            &fragment_stream,
+                            true,
+                            &mut validation_log,
+                        )
+                        .unwrap();
+
+                        let errors = report_split_errors(validation_log.get_log_mut());
+                        assert!(errors.is_empty());
+                    }
+
+                    // test verifying all at once
+                    let mut output_fragments = Vec::new();
+                    for entry in &fragments {
+                        output_fragments.push(new_output_path.join(entry.file_name().unwrap()));
+                    }
+
+                    let mut reader = Cursor::new(init_stream);
+                    let mut validation_log = DetailedStatusTracker::new();
+                    let _manifest = Store::load_from_file_and_fragments(
+                        "mp4",
+                        &mut reader,
+                        &output_fragments,
+                        true,
+                        &mut validation_log,
+                    )
+                    .unwrap();
+
+                    let errors = report_split_errors(validation_log.get_log_mut());
+                    assert!(errors.is_empty());
+                }
+                Err(_) => panic!("test misconfigures"),
+            }
+        }
     }
 }