+++ sdk/src/utils/mod.rs
@@ -15,10 +15,12 @@ pub(crate) mod base64;
 pub(crate) mod cbor_types;
 #[allow(dead_code)]
 pub(crate) mod hash_utils;
+pub(crate) mod io_utils;
 pub(crate) mod merkle;
 pub(crate) mod mime;
 #[allow(dead_code)] // for wasm build
 pub(crate) mod patch;
+pub(crate) mod sig_utils;
 #[cfg(feature = "add_thumbnails")]
 pub(crate) mod thumbnail;
 pub(crate) mod time_it;